refactor: use identities instead of duplicating logic
