+++ packages/crypto/src/crypto/index.ts
@@ -1,8 +1,8 @@
 import * as bip38 from "./bip38";
-import { crypto } from "./crypto";
-import { HashAlgorithms } from "./hash-algorithms";
-import { HDWallet } from "./hdwallet";
-import { Message } from "./message";
-import { slots } from "./slots";
 
-export { crypto, HDWallet, Message, slots, HashAlgorithms, bip38 };
+export { bip38 }
+export { crypto } from "./crypto";
+export { HashAlgorithms } from "./hash-algorithms";
+export { HDWallet } from "./hdwallet";
+export { Message } from "./message";
+export { slots } from "./slots";
