+++ packages/crypto/src/identities/index.ts
@@ -1,5 +1,5 @@
 export { Address } from "./address";
-export { Keys } from "./keys";
+export { Keys, KeyPair } from "./keys";
 export { PrivateKey } from "./private-key";
 export { PublicKey } from "./public-key";
 export { WIF } from "./wif";
