+++ packages/crypto/src/identities/wif.ts
@@ -1,7 +1,7 @@
 import wif from "wif";
 import { configManager } from "../managers";
 import { INetwork } from "../networks";
-import { Keys } from "./keys";
+import { KeyPair, Keys } from "./keys";
 
 export class WIF {
     public static fromPassphrase(passphrase: string, network?: INetwork): string {
@@ -9,6 +9,14 @@ export class WIF {
 
         if (!network) {
             network = configManager.all();
+        }
+
+        return wif.encode(network.wif, Buffer.from(keys.privateKey, "hex"), keys.compressed);
+    }
+
+    public static fromKeys(keys: KeyPair, network?: INetwork): string {
+        if (!network) {
+            network = configManager.all() as INetwork;
         }
 
         return wif.encode(network.wif, Buffer.from(keys.privateKey, "hex"), keys.compressed);
